requirements.txt
playwright
chromadb
openai
openai
python-dotenv
